benchmarks: split individual test funcs out of benchmark.c framework
The idea here is to make adding/removing benchmark tests easier by
having benchmark.c be the framework and all the functioned needed
for individual benchmarks live in benches.c.
